Reorganize the tests class
